refactor: move TriggerStrategy to shared package
