+++ packages/pieces/calendly/src/lib/trigger/invitee-canceled.trigger.ts
@@ -1,4 +1,5 @@
-import { createTrigger, TriggerStrategy, HttpRequest, HttpMethod, AuthenticationType, httpClient } from "@activepieces/framework";
+import { createTrigger, HttpRequest, HttpMethod, AuthenticationType, httpClient } from "@activepieces/framework";
+import { TriggerStrategy } from "@activepieces/shared";
 import { calendlyCommon, CalendlyWebhookInformation } from "../common";
