+++ packages/pieces/google-sheets/src/lib/triggers/new-row-added.ts
@@ -1,4 +1,5 @@
-import { createTrigger, TriggerStrategy } from '@activepieces/framework';
+import { createTrigger } from '@activepieces/framework';
+import { TriggerStrategy } from '@activepieces/shared';
 import { googleSheetsCommon } from '../common/common';
 
 export const newRowAdded = createTrigger({
