+++ packages/pieces/pipedrive/src/lib/trigger/new-deal.ts
@@ -1,4 +1,5 @@
-import { createTrigger, TriggerStrategy } from '@activepieces/framework';
+import { createTrigger } from '@activepieces/framework';
+import { TriggerStrategy } from '@activepieces/shared';
 import { pipedriveCommon } from '../common';
 
 export const newDeal = createTrigger({
