+++ packages/pieces/slack/src/lib/triggers/new-message.ts
@@ -1,4 +1,5 @@
-import { createTrigger, TriggerStrategy } from "@activepieces/framework";
+import { createTrigger } from "@activepieces/framework";
+import { TriggerStrategy } from "@activepieces/shared";
 import { slackAuth, slackChannel } from "../common/props";
