+++ packages/pieces/todoist/src/lib/triggers/task-completed-trigger.ts
@@ -1,4 +1,5 @@
-import { createTrigger, TriggerStrategy, assertNotNullOrUndefined } from '@activepieces/framework';
+import { createTrigger, assertNotNullOrUndefined } from '@activepieces/framework';
+import { TriggerStrategy } from '@activepieces/shared';
 import dayjs from 'dayjs';
 import { todoistSyncClient } from '../common/client/sync-client'
 import { TodoistCompletedTask } from '../common/models'
