+++ packages/pieces/youtube/src/lib/triggers/new-video.trigger.ts
@@ -1,4 +1,5 @@
-import { createTrigger, TriggerStrategy, httpClient, HttpMethod } from '@activepieces/framework';
+import { createTrigger, httpClient, HttpMethod } from '@activepieces/framework';
+import { TriggerStrategy } from '@activepieces/shared';
 import { channelIdentifier } from '../common/props';
 import dayjs from 'dayjs';
 import cheerio from "cheerio";
